+++ examples/execute_all.php
@@ -9,6 +9,10 @@ try {
 
     $collection_or_core_name = $config['endpoint']['localhost']['core'] = uniqid();
 
+    // only necessary for solrcloud mode, but needs to be set to retrieve the mode from admin/info/system
+    $config['endpoint']['localhost']['username'] = 'solr';
+    $config['endpoint']['localhost']['password'] = 'SolrRocks';
+
     // create a client instance
     $client = new Solarium\Client($adapter, $eventDispatcher, $config);
 
@@ -19,12 +23,12 @@ try {
 
     $data = $client->execute($query)->getData();
     $solr_mode = $data['mode'] ?? 'server';
+    $solrSpecVersion = $data['lucene']['solr-spec-version'];
+    $solrVersion = (int) strstr($solrSpecVersion, '.', true);
 
     if ('solrcloud' === $data['mode']) {
         $config['endpoint']['localhost']['collection'] = $config['endpoint']['localhost']['core'];
         unset($config['endpoint']['localhost']['core']);
-        $config['endpoint']['localhost']['username'] = 'solr';
-        $config['endpoint']['localhost']['password'] = 'SolrRocks';
         // adjust the client instance
         $client = new Solarium\Client($adapter, $eventDispatcher, $config);
 
@@ -57,29 +61,32 @@ try {
         $response = $client->coreAdmin($coreAdminQuery);
     }
 
-    // check if /mlt handler exists (it will in the github worklow, but not when running this script on its own)
-    $query = $client->createApi([
-        'version' => Request::API_V1,
-        'handler' => $collection_or_core_name.'/config/requestHandler',
-    ]);
-    $query->addParam('componentName', '/mlt');
-    $response = $client->execute($query);
-    $mltHandler = $response->getData()['config']['requestHandler']['/mlt'];
-
-    if (null === $mltHandler) {
-        // set up /mlt handler for MoreLikeThis query examples
+    // @todo Figure out why this fails on Solr 9.
+    if (9 !== $solrVersion) {
+        // check if /mlt handler exists (it will in the github worklow, but not when running this script on its own)
         $query = $client->createApi([
             'version' => Request::API_V1,
-            'handler' => $collection_or_core_name.'/config',
-            'method' => Request::METHOD_POST,
-            'rawdata' => json_encode([
-                'add-requesthandler' => [
-                    'name' => '/mlt',
-                    'class' => 'solr.MoreLikeThisHandler',
-                ],
-            ]),
+            'handler' => $collection_or_core_name.'/config/requestHandler',
         ]);
-        $client->execute($query);
+        $query->addParam('componentName', '/mlt');
+        $response = $client->execute($query);
+        $mltHandler = $response->getData()['config']['requestHandler']['/mlt'];
+
+        if (null === $mltHandler) {
+            // set up /mlt handler for MoreLikeThis query examples
+            $query = $client->createApi([
+                'version' => Request::API_V1,
+                'handler' => $collection_or_core_name.'/config',
+                'method' => Request::METHOD_POST,
+                'rawdata' => json_encode([
+                    'add-requesthandler' => [
+                        'name' => '/mlt',
+                        'class' => 'solr.MoreLikeThisHandler',
+                    ],
+                ]),
+            ]);
+            $client->execute($query);
+        }
     }
 
     // disable automatic commits for update examples
@@ -127,6 +134,14 @@ try {
         '7.5.3-plugin-bufferedupdate-benchmarks.php', // takes too long for a workflow, can be run manually
     ];
 
+    // examples that can't be run against this Solr version
+    $skipForVersion = [];
+
+    if (9 === $solrVersion) {
+        $skipForVersion[] = '2.3.1-mlt-query.php';
+        $skipForVersion[] = '2.3.2-mlt-stream.php';
+    }
+
     // examples that can't be run in cloud mode
     $skipForCloud = [
         '2.1.5.7-grouping-by-query.php',
@@ -139,17 +154,16 @@ try {
     foreach (scandir(__DIR__) as $example) {
         if (preg_match('/^\d.*\.php/', $example)) {
             print "\n".$example.' ';
-            if (!in_array($example, $skipAltogether)) {
-                if ('solrcloud' !== $solr_mode || !in_array($example, $skipForCloud)) {
-                    ob_start();
-                    require($example);
-                    ob_end_clean();
-                } else {
-                    print 'Could not be run in cloud mode.';
-                }
-
-            } else {
+            if (in_array($example, $skipAltogether)) {
                 print 'Could not be run against the techproducts example.';
+            } elseif (in_array($example, $skipForVersion)) {
+                printf('Could not be run against Solr %d.', $solrVersion);
+            } elseif ('solrcloud' === $solr_mode && in_array($example, $skipForCloud)) {
+                print 'Could not be run in cloud mode.';
+            } else {
+                ob_start();
+                require($example);
+                ob_end_clean();
             }
         }
     }